fix: Don't scroll when a link scrolls and highlights text
Should fix, but requires confirmation: #10, #32, #46 and #51

When opening a file from wikilink/link with header and link with block id
(i.e. [[note#header]], [[note#^blockId]])
obsidian highlights and scrolls to the corresponding header and text.
The issue is that the plugin restores the saved scroll, despite the highlighted text.

A fix was implemented in 7bae48c by checking if obsidian has scrolled or not (`scroll === 0`). But, this doesn't work for files with properties: getScroll returns the cursor position, which is at the end of the properties when opening a file. For files with peroperties, scroll is not 0, so the plugin doesn't restore the state.
`scroll === 0` was removed in commits 768e992 and 8163776, waiting for another solution.

This new solution checks for any dom element in a span.is-flashing.
This dom is created by obsidian when a specific text is highlighted after clicking a link.

Since the check for `scroll === 0` was removed, the for loop shouldn't be necessary (introduced in 7bae48c).
delayAfterFileOpening too, but it might be worth keeping.